expose datacmd and dataquery
